Rename the the dialect directory to follow upstream convention.
