+++ mlir/include/mlir-extensions/Dialect/plier_util/dialect.hpp
@@ -1,4 +1,4 @@
-// Copyright 2022 Intel Corporation
+// Copyright 2021 Intel Corporation
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -24,21 +24,16 @@
 #include <mlir/Interfaces/SideEffectInterfaces.h>
 #include <mlir/Interfaces/ViewLikeInterface.h>
 
-#include <mlir/Dialect/GPU/GPUDialect.h>
-
-#include "mlir-extensions/dialect/gpu_runtime/IR/GpuRuntimeOpsDialect.h.inc"
-
-#include "mlir-extensions/dialect/gpu_runtime/IR/GpuRuntimeOpsEnums.h.inc"
+#include "mlir-extensions/Dialect/plier_util/PlierUtilOpsDialect.h.inc"
+#include "mlir-extensions/Dialect/plier_util/PlierUtilOpsEnums.h.inc"
 
 #define GET_ATTRDEF_CLASSES
-#include "mlir-extensions/dialect/gpu_runtime/IR/GpuRuntimeOpsAttributes.h.inc"
+#include "mlir-extensions/Dialect/plier_util/PlierUtilOpsAttributes.h.inc"
 
 #define GET_OP_CLASSES
-#include "mlir-extensions/dialect/gpu_runtime/IR/GpuRuntimeOps.h.inc"
+#include "mlir-extensions/Dialect/plier_util/PlierUtilOps.h.inc"
 
-// TODO: OpaqueType is not tied to gpu dialect. Move this or create separate
-// types
-namespace gpu_runtime {
+namespace plier {
 class OpaqueType : public ::mlir::Type::TypeBase<OpaqueType, ::mlir::Type,
                                                  ::mlir::TypeStorage> {
 public:
@@ -47,7 +42,4 @@ public:
   static OpaqueType get(mlir::MLIRContext *context);
 };
 
-mlir::StringRef getAllocSharedAttrName();
-mlir::StringRef getGpuAccessibleAttrName();
-
-} // namespace gpu_runtime
+} // namespace plier
