+++ mlir/include/mlir-extensions/transforms/call_lowering.hpp
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "mlir-extensions/dialect/plier/dialect.hpp"
+#include "mlir-extensions/Dialect/plier/dialect.hpp"
 
 #include <mlir/IR/PatternMatch.h>
