+++ mlir/include/mlir-extensions/transforms/cast_lowering.hpp
@@ -16,7 +16,7 @@
 
 #include <functional>
 
-#include "mlir-extensions/dialect/plier/dialect.hpp"
+#include "mlir-extensions/Dialect/plier/dialect.hpp"
 
 #include <mlir/IR/PatternMatch.h>
