+++ mlir/lib/Conversion/gpu_runtime_to_llvm.cpp
@@ -14,8 +14,8 @@
 
 #include "mlir-extensions/Conversion/gpu_runtime_to_llvm.hpp"
 
-#include "mlir-extensions/dialect/gpu_runtime/IR/gpu_runtime_ops.hpp"
-#include "mlir-extensions/dialect/plier_util/dialect.hpp"
+#include "mlir-extensions/Dialect/gpu_runtime/IR/gpu_runtime_ops.hpp"
+#include "mlir-extensions/Dialect/plier_util/dialect.hpp"
 #include "mlir-extensions/transforms/func_utils.hpp"
 
 #include <mlir/Conversion/AsyncToLLVM/AsyncToLLVM.h>
