+++ mlir/lib/transforms/promote_to_parallel.cpp
@@ -14,7 +14,7 @@
 
 #include "mlir-extensions/transforms/promote_to_parallel.hpp"
 
-#include "mlir-extensions/dialect/plier/dialect.hpp"
+#include "mlir-extensions/Dialect/plier/dialect.hpp"
 #include <mlir/Dialect/SCF/SCF.h>
 #include <mlir/IR/BlockAndValueMapping.h>
 #include <mlir/Interfaces/CallInterfaces.h>
